UI: Fixed download gateway launch file, clear code and refactoring downloads file
