+++ ui-ngx/src/app/shared/models/device.models.ts
@@ -865,14 +865,6 @@ export interface PublishTelemetryCommand {
   snmp?: string;
 }
 
-export interface PublishLaunchCommand {
-  mqtt: {
-    linux: string;
-    windows: string;
-    macos: string;
-  };
-}
-
 export const dayOfWeekTranslations = new Array<string>(
   'device-profile.schedule-day.monday',
   'device-profile.schedule-day.tuesday',
